@@ -1,5 +1,5 @@
 package com.bignardranch.android.geoquiz
 
-import android.support.annotation.StringRes
+import androidx.annotation.StringRes
 
 data class Question(@StringRes val textResId: Int, val answer: Boolean)
