@@ -1,22 +1,22 @@
 package com.bignardranch.android
 
+//import android.content.res.Configuration.ORIENTATION_LANDSCAPE
 import android.os.Bundle
+import android.util.Log
 import android.view.View
 import android.widget.Button
 import android.widget.TextView
 import android.widget.Toast
-import androidx.activity.viewModels // ??
+import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
+import androidx.core.os.bundleOf
+
 import com.bignardranch.android.geoquiz.GeoQuizViewModel
-import com.bignardranch.android.geoquiz.Question
 
 
 private const val TAG = "MainActivity"
-private const val DEBUG = false
+const val DEBUG = true
 
-/**
- * @see android.widget.Toast.setGravity
- */
 class MainActivity : AppCompatActivity() {
 
     private lateinit var questionTextView: TextView
@@ -24,26 +24,12 @@ class MainActivity : AppCompatActivity() {
     private lateinit var falseButton: Button
     private lateinit var prevButton: View // ImageButton in activity_main.xml and Button in land-activity_main.xml
     private lateinit var nextButton: View //
-    private val model: GeoQuizViewModel by viewModels() // NOT WORKING
 
-    private val questions = listOf(
-        Question(R.string.question_australia, true),
-        Question(R.string.question_oceans, true),
-        Question(R.string.question_mideast, false),
-        Question(R.string.question_africa, false),
-        Question(R.string.question_americas, true),
-        Question(R.string.question_asia, true)
-    )
+    private val viewModel: GeoQuizViewModel by viewModels() // TODO: see delegates
 
-    private val checkedAnswers = MutableList(6) { false }
-
-    private var currentQuestion = 0
-    private var correctAnswers = 0
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
-        //val model: GeoQuizViewModel by activityViewModels() // NOT WORKING
 
         if (DEBUG) {
             Toast.makeText(
@@ -52,8 +38,14 @@ class MainActivity : AppCompatActivity() {
                 Toast.LENGTH_SHORT
             ).show()
         }
-
+        Log.d(TAG, "onCreate() called")
         setContentView(R.layout.activity_main)
+
+        savedInstanceState?.let {
+            viewModel.currentQuestion = it.getInt("currentQuestion", 0)
+            viewModel.correctAnswers = it.getInt("currentAnswers", 0)
+            viewModel.checkedAnswers = it.getBooleanArray("checkedAnswers")?.toMutableList()!!
+        }
 
         questionTextView = findViewById(R.id.question_text_view)
         trueButton = findViewById(R.id.true_button)
@@ -84,6 +76,77 @@ class MainActivity : AppCompatActivity() {
         updateQuestion()
     }
 
+    private fun showNextQuestion() {
+        with(viewModel) { currentQuestion = ++currentQuestion % questionsSize }
+        updateQuestion()
+    }
+
+    private fun showPrevQuestion() {
+        with(viewModel) {
+            --currentQuestion
+            if (currentQuestion < 0) currentQuestion = questionsSize - 1
+        }
+        updateQuestion()
+    }
+
+    private fun updateQuestion() {
+        questionTextView.setText(viewModel.currentQuestionText)
+        if (viewModel.isCurrentAnswerChecked()) setAnswerButtonsEnabled(false)
+        else setAnswerButtonsEnabled(true)
+    }
+
+    private fun checkAnswer(answer: Boolean) {
+        setAnswerButtonsEnabled(false)
+
+        with(viewModel) {
+            val toastStr = if (answer == currentQuestionAnswer) {
+                ++correctAnswers
+                R.string.correct_toast
+            } else R.string.incorrect_toast
+
+            setCurrentCheckedAnswer()
+
+            val toast = if (isAllAnswersChecked()) {
+                Toast.makeText(
+                    this@MainActivity,
+                    String.format(
+                        "Correct! Your have answered $correctAnswers of $questionsSize (%.2f%%)",
+                        correctAnswers.toDouble() / questionsSize.toDouble() * 100.toDouble()
+                    ),
+                    Toast.LENGTH_LONG
+                )
+            } else {
+                Toast.makeText(
+                    this@MainActivity,
+                    toastStr,
+                    Toast.LENGTH_SHORT
+                )
+            }
+//            toast.setGravity(
+//                if (resources.configuration.orientation == ORIENTATION_LANDSCAPE) Gravity.TOP else Gravity.BOTTOM,
+//                0, 0
+//            )
+            toast.show()
+        }
+    }
+
+    private fun setAnswerButtonsEnabled(enabled: Boolean) {
+        falseButton.isEnabled = enabled
+        trueButton.isEnabled = enabled
+    }
+
+    override fun onSaveInstanceState(savedInstanceState: Bundle) {
+        super.onSaveInstanceState(savedInstanceState)
+        Log.d(TAG, "onSaveInstanceState() called")
+        savedInstanceState.putAll(
+            bundleOf(                                   // TODO: use delegates for keys
+                "currentQuestion" to viewModel.currentQuestion,
+                "correctAnswers" to viewModel.correctAnswers,
+                "checkedAnswers" to viewModel.checkedAnswers.toBooleanArray()
+            )
+        )
+    }
+
     override fun onStart() {
         super.onStart()
         if (DEBUG) {
@@ -93,6 +156,7 @@ class MainActivity : AppCompatActivity() {
                 Toast.LENGTH_SHORT
             ).show()
         }
+        Log.d(TAG, "onStart() called")
     }
 
     override fun onResume() {
@@ -104,6 +168,7 @@ class MainActivity : AppCompatActivity() {
                 Toast.LENGTH_SHORT
             ).show()
         }
+        Log.d(TAG, "onResume() called")
     }
 
     override fun onPause() {
@@ -111,10 +176,11 @@ class MainActivity : AppCompatActivity() {
         if (DEBUG) {
             Toast.makeText(
                 this,
-                "onPause() Toast",
+                "onPause() Toast, isFinishing == $isFinishing",
                 Toast.LENGTH_SHORT
             ).show()
         }
+        Log.d(TAG, "onPause() called, isFinishing == $isFinishing")
     }
 
     override fun onStop() {
@@ -126,6 +192,7 @@ class MainActivity : AppCompatActivity() {
                 Toast.LENGTH_SHORT
             ).show()
         }
+        Log.d(TAG, "onStop() called")
     }
 
     override fun onDestroy() {
@@ -137,52 +204,6 @@ class MainActivity : AppCompatActivity() {
                 Toast.LENGTH_SHORT
             ).show()
         }
-    }
-
-    private fun showNextQuestion() {
-        currentQuestion = ++currentQuestion % questions.size
-        updateQuestion()
-    }
-
-    private fun showPrevQuestion() {
-        --currentQuestion
-        if (currentQuestion < 0) currentQuestion = questions.size - 1
-        updateQuestion()
-    }
-
-    private fun updateQuestion() {
-        questionTextView.setText(questions[currentQuestion].textResId)
-        if (checkedAnswers[currentQuestion]) setAnswerButtonsEnabled(false)
-        else setAnswerButtonsEnabled(true)
-    }
-
-    private fun checkAnswer(answer: Boolean) {
-        setAnswerButtonsEnabled(false)
-
-        val toastStr = if (answer == questions[currentQuestion].answer) {
-            ++correctAnswers
-            R.string.correct_toast
-        } else R.string.incorrect_toast
-
-        checkedAnswers[currentQuestion] = true
-
-        val toast = if (checkedAnswers.all { it }) {
-            Toast.makeText(
-                this,
-                String.format("Correct! Your have answered $correctAnswers of ${questions.size} (%.2f%%)", correctAnswers.toDouble() / questions.size.toDouble() * 100.toDouble()),
-                Toast.LENGTH_LONG
-            )
-        } else {
-            Toast.makeText(
-                this,
-                toastStr,
-                Toast.LENGTH_SHORT
-            )
-        }.show()
-    }
-
-    private fun setAnswerButtonsEnabled(enabled: Boolean) {
-        falseButton.isEnabled = enabled
-        trueButton.isEnabled = enabled
+        Log.d(TAG, "onDestroy() called")
     }
 }
